Move Card to Page Sections
