+++ src/Pages/Sections/Card/PublicPredictions.tsx
@@ -1,7 +1,7 @@
 import { Grid, Typography } from '@material-ui/core';
 import React, { ReactElement } from 'react';
 
-import { IFixtureWithScoreAndPredictions } from '../../models/toto.models';
+import { IFixtureWithScoreAndPredictions } from '../../../models/toto.models';
 
 type Props = { fixtureWithPrediction: IFixtureWithScoreAndPredictions };
